requirements.txt
torch==2.2.2
torchvision==0.17.2
facenet-pytorch
numpy>=1.24.0,<2.0.0
opencv-contrib-python==4.8.1.78
Pillow
timm
psutil
matplotlib
seaborn
pandas
tabulate
